refactor: Remove duplicate global color definition
