--- @navikt/core/tokens/darkside/tokens/global/accent.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const AccentScaleLight = () => getGlobalScaleForColor("accent", "light");
-const AccentScaleDark = () => getGlobalScaleForColor("accent", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"accent"> => {
-  return {
-    accent: theme === "light" ? AccentScaleLight() : AccentScaleDark(),
-  };
-};
