--- @navikt/core/tokens/darkside/tokens/global/brand-one.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const BrandOneScaleLight = () => getGlobalScaleForColor("brandOne", "light");
-const BrandOneScaleDark = () => getGlobalScaleForColor("brandOne", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"brandOne"> => {
-  return {
-    brandOne: theme === "light" ? BrandOneScaleLight() : BrandOneScaleDark(),
-  };
-};
