--- @navikt/core/tokens/darkside/tokens/global/brand-three.ts
@@ -1,13 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const BrandThreeScaleLight = () =>
-  getGlobalScaleForColor("brandThree", "light");
-const BrandThreeScaleDark = () => getGlobalScaleForColor("brandThree", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"brandThree"> => {
-  return {
-    brandThree:
-      theme === "light" ? BrandThreeScaleLight() : BrandThreeScaleDark(),
-  };
-};
