--- @navikt/core/tokens/darkside/tokens/global/brand-two.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const BrandTwoScaleLight = () => getGlobalScaleForColor("brandTwo", "light");
-const BrandTwoScaleDark = () => getGlobalScaleForColor("brandTwo", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"brandTwo"> => {
-  return {
-    brandTwo: theme === "light" ? BrandTwoScaleLight() : BrandTwoScaleDark(),
-  };
-};
