--- @navikt/core/tokens/darkside/tokens/global/data-one.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const DataOneScaleLight = () => getGlobalScaleForColor("dataOne", "light");
-const DataOneScaleDark = () => getGlobalScaleForColor("dataOne", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"dataOne"> => {
-  return {
-    dataOne: theme === "light" ? DataOneScaleLight() : DataOneScaleDark(),
-  };
-};
