--- @navikt/core/tokens/darkside/tokens/global/data-two.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const DataTwoScaleLight = () => getGlobalScaleForColor("dataTwo", "light");
-const DataTwoScaleDark = () => getGlobalScaleForColor("dataTwo", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"dataTwo"> => {
-  return {
-    dataTwo: theme === "light" ? DataTwoScaleLight() : DataTwoScaleDark(),
-  };
-};
