--- @navikt/core/tokens/darkside/tokens/global/info.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const InfoScaleLight = () => getGlobalScaleForColor("info", "light");
-const InfoScaleDark = () => getGlobalScaleForColor("info", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"info"> => {
-  return {
-    info: theme === "light" ? InfoScaleLight() : InfoScaleDark(),
-  };
-};
