--- @navikt/core/tokens/darkside/tokens/global/success.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const SuccessScaleLight = () => getGlobalScaleForColor("success", "light");
-const SuccessScaleDark = () => getGlobalScaleForColor("success", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"success"> => {
-  return {
-    success: theme === "light" ? SuccessScaleLight() : SuccessScaleDark(),
-  };
-};
