--- @navikt/core/tokens/darkside/tokens/global/warning.ts
@@ -1,11 +1,0 @@
-import { getGlobalScaleForColor } from "../../leonardo";
-import { ColorThemeMode, GlobaColorScale } from "../../util";
-
-const WarningScaleLight = () => getGlobalScaleForColor("warning", "light");
-const WarningScaleDark = () => getGlobalScaleForColor("warning", "dark");
-
-export default (theme: ColorThemeMode): GlobaColorScale<"warning"> => {
-  return {
-    warning: theme === "light" ? WarningScaleLight() : WarningScaleDark(),
-  };
-};
